Control.Monad.Prompt.Class

Description

Provides a typeclass for Applicative and Monad types that give you the ability to, at any time, "prompt" with an a and get a b in response. The power of this instance is that each type gets to define its own way to deliver a response.

This library provides instances for PromptT from Control.Monad.Prompt and the monad transformers in transformers and mtl. Feel free to create your own instances too.

data Interactive a = Interactive ((String -> String) -> a)

-- at any time, ask with a string to get a string
instance MonadPrompt String String Interactive where
    prompt str = Interactive $ f -> f str

Documentation

class Applicative m => MonadPrompt a b m | m -> a b where

An Applicative (and possibly Monad) where you can, at any time, "prompt" with an a and receive a b in response.

Instances include PromptT and any transformers monad transformer over another MonadPrompt.

Minimal complete definition

prompt | prompts

Methods

prompt :: a -> m b

Prompt with an a for a b in the context of the type.

prompts :: (b -> c) -> a -> m c

Prompt with an a for a b in the context of the type, and apply the given function to receive a c.

prompt' :: MonadPrompt a b m => a -> m b

A version of prompt strict on its prompting value.

prompts' :: MonadPrompt a b m => (b -> c) -> a -> m c

A version of prompts strict on its prompting value.